Embedding functions

We provide various functions with which to compute SVD embeddings.

Detailed API

​

DotProductGraphs.dot_product — Function

dotproduct(L,R;toprob = true)

Computes the dot product between two embeddings and gives back a matrix of interaction probabilities

The left and right embeddings must be of the same dimension (e.g., size(L)[2] == size(R)[2]) but can be of different number of nodes (e.g. for bipartite networks).

Arguments

L: Left embedding

R: Right embedding

to_prob: (optional) whether to clamp the outcomes to the interval [0,1]

Examples

julia> # we first build a 2 blocks matrix:
julia> block_matrix = reshape([ones(5,5) zeros(5,5); zeros(5,5) ones(5,5)],10,10)
julia> # and then decompose it:
julia> L,R = svd_embedding(block_matrix,2)
julia> dot_product(L,R) ≈ block_matrix

DotProductGraphs.svd_embedding — Function

svdembedding(A,d; svdengine = nothing)

Computes an SVD embedding of an adjacency matrix A of dimension d, using svd_engine to perform the SVD factorization

Given an adjacency matrix A, the function returns its node embedding using the Singular Value Decomposition, as it is done in Random Dot Product Graphs. The functions accepts a user defined svd_engine that can be different from the detaul svd() coming from LinearAlgebra. In that case, svd_engine must be a function of the same form of truncated_svd.

Arguments

A: Adjacency matrix of the graph to embed.

d: Dimension of the embedding

svd_engine: The function used to perform the SVD factorization, by default svd() from LinearAlgebra

Notes

The matrix A can be asymmetric (that is, the graph can be directed) and rectangular (e.g., for bipartite graphs).

Examples

julia> # we first build a 2 blocks matrix:
julia> block_matrix = reshape([ones(5,5) zeros(5,5); zeros(5,5) ones(5,5)],10,10)
julia> # and then decompose it:
julia> L,R = svd_embedding(block_matrix,2)

DotProductGraphs.truncated_svd — Function

truncated_svd(Mat, dim)

Computes a truncated Singular Value Decomposition

Just some rice over the default LinearAlgebra.svd to return a truncated Singular Value Decomposition of dimension d.

Arguments

Mat: The matrix we want to factorize

d: The dimension of the factorization

Notes

This is in general not efficient, as we are computing first a full factorization and then truncating it. We should consider using more optimized methods.

Examples

julia> # we first build a 2 blocks matrix:
julia> block_matrix = reshape([ones(5,5) zeros(5,5); zeros(5,5) ones(5,5)],10,10)
julia> # and then decompose it:
julia> L, Σ, Rt = truncated_svd(block_matrix,2)